rest-full/RestApi/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
asn1crypto==1.4.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-rest-auth==2.1.4
Django==3.2.2
django-allauth==0.44.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
duplicity==0.8.17
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
jeepney==0.6.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mechanize==0.4.5
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
pbr==5.6.0
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-apt==2.1.7+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.39
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.4.0+ds
wadllib==1.3.5
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
